[MIG] shopinvader_sale_packaging + split into shopinvader_product_sale_packaging and shopinvader_restapi_sale_packaging
