+++ setup/shopinvader_product_sale_packaging/odoo/addons/shopinvader_product_sale_packaging
@@ -1,0 +1,1 @@
+../../../../shopinvader_product_sale_packaging
