+++ setup/shopinvader_product_sale_packaging/setup.py
