+++ setup/shopinvader_restapi_sale_packaging/odoo/addons/shopinvader_restapi_sale_packaging
@@ -1,0 +1,1 @@
+../../../../shopinvader_restapi_sale_packaging
