--- setup/shopinvader_sale_packaging/odoo/addons/shopinvader_sale_packaging
@@ -1,1 +1,0 @@
-../../../../shopinvader_sale_packaging
