+++ shopinvader_product_sale_packaging/__init__.py
@@ -1,0 +1,2 @@
+from . import models
+from . import schemas
