+++ shopinvader_product_sale_packaging/models/__init__.py
@@ -1,0 +1,3 @@
+from . import product_packaging
+from . import product_packaging_level
+from . import product_product
