+++ shopinvader_product_sale_packaging/readme/CONTRIBUTORS.rst
@@ -1,0 +1,2 @@
+* Simone Orsi <simahawk@gmail.com>
+* Marie Lejeune <marie.lejeune@acsone.eu>
