+++ shopinvader_product_sale_packaging/readme/CREDITS.rst
