+++ shopinvader_product_sale_packaging/readme/DESCRIPTION.rst
@@ -1,0 +1,3 @@
+Provide packaging information for products.
+
+Available packaging are taken using `stock_packaging_calculator <https://github.com/OCA/stock-logistics-warehouse/tree/16.0/stock_packaging_calculator>`_.
