+++ shopinvader_product_sale_packaging/schemas/__init__.py
@@ -1,0 +1,1 @@
+from . import product
