+++ shopinvader_product_sale_packaging/views/product_packaging_level.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+    <record id="view_product_packaging_level_form" model="ir.ui.view">
+        <field name="model">product.packaging.level</field>
+        <field
+            name="inherit_id"
+            ref="product_packaging_level_sellable.view_product_packaging_level_form"
+        />
+        <field name="arch" type="xml">
+            <field name="can_be_sold" position="after">
+                <field name="shopinvader_display" />
+            </field>
+        </field>
+    </record>
+</odoo>
