+++ shopinvader_restapi_sale_packaging/readme/CREDITS.rst
@@ -1,0 +1,4 @@
+The development of this module has been financially supported by:
+
+* Camptocamp
+* Cosanum
