+++ shopinvader_restapi_sale_packaging/readme/DESCRIPTION.rst
@@ -1,0 +1,3 @@
+Allow to purchase products by packaging.
+
+Available packaging are taken using `stock_packaging_calculator <https://github.com/OCA/stock-logistics-warehouse/tree/16.0/stock_packaging_calculator>`_.
