--- shopinvader_sale_packaging/__init__.py
@@ -1,2 +1,0 @@
-from . import services
-from . import models
