--- shopinvader_sale_packaging/i18n/shopinvader_sale_packaging.pot
@@ -1,72 +1,0 @@
-# Translation of Odoo Server.
-# This file contains the translation of the following modules:
-# 	* shopinvader_sale_packaging
-#
-msgid ""
-msgstr ""
-"Project-Id-Version: Odoo Server 14.0\n"
-"Report-Msgid-Bugs-To: \n"
-"Last-Translator: \n"
-"Language-Team: \n"
-"MIME-Version: 1.0\n"
-"Content-Type: text/plain; charset=UTF-8\n"
-"Content-Transfer-Encoding: \n"
-"Plural-Forms: \n"
-
-#. module: shopinvader_sale_packaging
-#: model:ir.model.fields,field_description:shopinvader_sale_packaging.field_product_packaging__display_name
-#: model:ir.model.fields,field_description:shopinvader_sale_packaging.field_product_packaging_type__display_name
-#: model:ir.model.fields,field_description:shopinvader_sale_packaging.field_shopinvader_variant__display_name
-msgid "Display Name"
-msgstr ""
-
-#. module: shopinvader_sale_packaging
-#: model:ir.model.fields,field_description:shopinvader_sale_packaging.field_product_packaging__id
-#: model:ir.model.fields,field_description:shopinvader_sale_packaging.field_product_packaging_type__id
-#: model:ir.model.fields,field_description:shopinvader_sale_packaging.field_shopinvader_variant__id
-msgid "ID"
-msgstr ""
-
-#. module: shopinvader_sale_packaging
-#: model:ir.model.fields,help:shopinvader_sale_packaging.field_product_packaging__shopinvader_display
-#: model:ir.model.fields,help:shopinvader_sale_packaging.field_product_packaging_type__shopinvader_display
-msgid "Include this packaging into Shopinvader product metadata."
-msgstr ""
-
-#. module: shopinvader_sale_packaging
-#: model:ir.model.fields,field_description:shopinvader_sale_packaging.field_product_packaging____last_update
-#: model:ir.model.fields,field_description:shopinvader_sale_packaging.field_product_packaging_type____last_update
-#: model:ir.model.fields,field_description:shopinvader_sale_packaging.field_shopinvader_variant____last_update
-msgid "Last Modified on"
-msgstr ""
-
-#. module: shopinvader_sale_packaging
-#: model:ir.model.fields,field_description:shopinvader_sale_packaging.field_shopinvader_variant__packaging
-msgid "Packaging"
-msgstr ""
-
-#. module: shopinvader_sale_packaging
-#: model:ir.model,name:shopinvader_sale_packaging.model_product_packaging
-msgid "Product Packaging"
-msgstr ""
-
-#. module: shopinvader_sale_packaging
-#: model:ir.model.fields,field_description:shopinvader_sale_packaging.field_product_packaging__shopinvader_display
-#: model:ir.model.fields,field_description:shopinvader_sale_packaging.field_product_packaging_type__shopinvader_display
-msgid "Shopinvader Display"
-msgstr ""
-
-#. module: shopinvader_sale_packaging
-#: model:ir.model,name:shopinvader_sale_packaging.model_shopinvader_variant
-msgid "Shopinvader Variant"
-msgstr ""
-
-#. module: shopinvader_sale_packaging
-#: model:ir.model.fields,help:shopinvader_sale_packaging.field_shopinvader_variant__packaging
-msgid "Technical field to store packaging for the shop"
-msgstr ""
-
-#. module: shopinvader_sale_packaging
-#: model:ir.model,name:shopinvader_sale_packaging.model_product_packaging_type
-msgid "Type management for product.packaging"
-msgstr ""
