--- shopinvader_sale_packaging/models/__init__.py
@@ -1,3 +1,0 @@
-from . import shopinvader_variant
-from . import product_packaging
-from . import product_packaging_type
