--- shopinvader_sale_packaging/readme/CONTRIBUTORS.rst
@@ -1,1 +1,0 @@
-* Simone Orsi <simahawk@gmail.com>
