--- shopinvader_sale_packaging/readme/DESCRIPTION.rst
@@ -1,3 +1,0 @@
-Provide packaging information for products and allow to purchase them by packaging.
-
-Available packaging are taken using `stock_packaging_calculator <https://github.com/OCA/stock-logistics-warehouse/tree/13.0/stock_packaging_calculator>`_.
