--- shopinvader_sale_packaging/views/product_packaging_type.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-    <record id="view_product_packaging_type_form" model="ir.ui.view">
-        <field name="model">product.packaging.type</field>
-        <field
-            name="inherit_id"
-            ref="sale_by_packaging.view_product_packaging_type_form"
-        />
-        <field name="arch" type="xml">
-            <field name="can_be_sold" position="after">
-                <field name="shopinvader_display" />
-            </field>
-        </field>
-    </record>
-</odoo>
